Add Python binding compatibility with new hashing function
The new functions require libav which isn't statically
linked for the Python binding.

This allows the Python binding to use the old function. The new
hashing function is removed by the pre-processor using
#define CYTHON which is passed in to setup.py to the Makefile.

For the future it seems like it would be simpler to statically link it
for both. The Python binding aren't compatible with Windows
anyway, so portability doesn't seem like too much of a concern.